LayoutTests/imported/w3c/web-platform-tests/IndexedDB/structured-clone.any-expected.txt

CONSOLE MESSAGE: line 98: SyntaxError: No identifiers allowed directly after numeric literal

Harness Error (FAIL), message = SyntaxError: No identifiers allowed directly after numeric literal
